worker/requirements.txt
httpx==0.27.2
redis==5.0.8
orjson==3.10.7
langchain==0.3.3
langchain-core==0.3.3
langchain-openai==0.2.0
langgraph==0.2.16
openai==1.51.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
tenacity==8.5.0
SQLAlchemy==2.0.35
psycopg2-binary==2.9.9

# LLM Dependencies
transformers>=4.37.0
torch>=2.1.0
accelerate>=0.25.0


